Extract trajectory text sidecar codec
